Create SearchLoop.java
This is the loop that allows for CSV values to be found by establishing searchVal as the Column header and to return all data that would be under it. TO DO: Create method for testing the the next vertical cell down so that when the end of the column is reached, the loop ends.